python_analysis_server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
nltk==3.8.1
textblob==0.17.1
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0